% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts.R
\name{as_gts}
\alias{as_gts}
\title{Coersce object to GTS tibbles}
\usage{
as_gts(x, class = NULL, labels = list())
}
\arguments{
\item{x}{An object that could reasoably be coerced to a gts tibble.}

\item{class}{The class of the GTS.}

\item{labels}{A named list of labels in the form `key = value`.}
}
\description{
`as_gts()` turns an existing object, such as a data frame into
an extended tibble with class ['gts'], and attributes corresponding
to the GTS attributes + class.
A [`GTS`] object is basically a data frame with some information about the
timeserie.
The resulting tibble contains at least two columns: `timestamp` and `value`.

`as_gts()` is an S3 generic.
}
\examples{

x <- data.frame(timestamp = 1:10, value = rnorm(10))
as_gts(x)

}
