% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_endpoint}
\alias{assert_endpoint}
\title{Assert Endpoint}
\usage{
assert_endpoint(endpoint)
}
\arguments{
\item{endpoint}{A string defining the endpoint of the Warp 10 database.}
}
\description{
Check that endpoint is available.
}
\seealso{
Other assert: \code{\link{assert_token}}
}
\concept{assert}
