% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_token}
\alias{assert_token}
\title{Assert Token}
\usage{
assert_token(token)
}
\arguments{
\item{token}{A string defining the token used to fetch or update data.}
}
\description{
Check that token is available.
}
\seealso{
Other assert: 
\code{\link{assert_endpoint}()}
}
\concept{assert}
