% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_gts.R
\name{bind_lgts}
\alias{bind_lgts}
\title{Bind LGTS}
\usage{
bind_lgts(lgts, combine = FALSE, .funs = "first")
}
\arguments{
\item{lgts}{LGTS dataframe}

\item{combine}{If TRUE, combine multiple time series into one by grouping dates.}

\item{.funs}{A function \code{fun}, a quosure style lambda \code{~ fun(.)} or a list of either form.}
}
\description{
Bind LGTS into a single GTS
}
