% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{build_res}
\alias{build_res}
\alias{build_res.data}
\alias{build_res.ldata}
\alias{build_res.default}
\alias{build_res.map}
\alias{build_res.lastactivity}
\alias{build_res.list}
\alias{build_res.gts}
\alias{build_res.lgts}
\title{Build results}
\usage{
build_res(data, ...)

\method{build_res}{data}(data, ...)

\method{build_res}{ldata}(data, tz)

\method{build_res}{default}(data, ...)

\method{build_res}{map}(data, ...)

\method{build_res}{lastactivity}(data, tz)

\method{build_res}{list}(data, ...)

\method{build_res}{gts}(data, tz)

\method{build_res}{lgts}(data, tz)
}
\arguments{
\item{data}{A list resulting of a parsed json of all results.}

\item{...}{Other paramters passed into methods}

\item{tz}{Time Zone}
}
\description{
Build results from parsed json file.
}
