% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation}
\alias{documentation}
\title{Parameter documentation}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{class}{Selector for the GTS classes.
Can either be an exact match or a regular expression (the value should then start with a ~).}

\item{endpoint}{egress endpoint.}

\item{execute}{A boolean indicating if the warpScript should be exectuted or not.}

\item{labels}{Map of labels selectors.
The key is the label name, the value a selector which can be an exact match or a regular expression
(the value should then start with a ~).}

\item{attributes}{Map of attributes.}

\item{token}{a token needed to connect to database}
}
\description{
This function is only here to gather the documentation of different parameters used
throughout the package.
}
