% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{extract_gts}
\alias{extract_gts}
\title{Extract Geo Time Series}
\usage{
extract_gts(response, with_labels = FALSE)
}
\arguments{
\item{response}{response body of a post request (a character vector or a json)}

\item{with_labels}{if TRUE, column names also include Labels. Default to FALSE}
}
\value{
data.table
}
\description{
Extract all GTS from a JSON response and return them merged as a data.table.
If too slow, consider using postWarpscript with outputType="data.table" instead.
}
