% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{permalink}
\alias{permalink}
\title{Generate a Permalink}
\usage{
permalink(
  warpscript,
  plot = FALSE,
  endpoint = "http://localhost:8080/api/v0/exec",
  quantum = "http://localhost:8090"
)
}
\arguments{
\item{warpscript}{code or file name ending with .mc2}

\item{plot}{if TRUE, the generated link points to a plot. Default to FALSE}

\item{endpoint}{egress endpoint. Default to "http://localhost:8080/api/v0/exec"}

\item{quantum}{address of quantum instance. Default to "http://localhost:8090"}
}
\value{
url
}
\description{
Generate a permalink to a quantum instance.
}
