% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{post_warpscript}
\alias{post_warpscript}
\title{Post Warpscript Code}
\usage{
post_warpscript(warpscript, endpoint = get_endpoint())
}
\arguments{
\item{warpscript}{code or file name ending with .mc2}

\item{endpoint}{egress endpoint.}
}
\value{
json
}
\description{
Post warpscript code to a Warp 10 instance and retrieve response as character vector, json,
named list or data.table.
If outputType is "data.table", only the top of the stack is converted and it must be a list of GTS.
If this list is not a singleton, these GTS must have at most one value per timestamp.
}
