% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{postWarpscript}
\alias{postWarpscript}
\title{Post Warpscript Code}
\usage{
postWarpscript(warpscript, outputType = "json",
  endpoint = "http://localhost:8080/api/v0/exec", withLabels = FALSE)
}
\arguments{
\item{warpscript}{code or file name ending with .mc2}

\item{outputType}{the type of the returned value. The supported types are "raw", "json", "pretty", "list" and "data.table". Default to "json". Output type "dataFrame" is deprecated in favor of "data.table".}

\item{endpoint}{egress endpoint. Default to "http://localhost:8080/api/v0/exec"}

\item{withLabels}{if TRUE and if outputType is "data.table", column names also include Labels. Default to FALSE}
}
\value{
character vector or json or named list or data.table
}
\description{
Post warpscript code to a Warp 10 instance and retrieve response as character vector, json, named list or data.table. If outputType is "data.table", only the top of the stack is converted and it must be a list of GTS. If this list is not a singleton, these GTS must have at most one value per timestamp.
}
