% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{pushWarp10}
\alias{pushWarp10}
\title{Push data points}
\usage{
pushWarp10(data, token, endpoint = "http://localhost:8080/api/v0/update")
}
\arguments{
\item{data}{data points in GTS input format as a character vector or a filename ending with .data or .gz}

\item{token}{write token}

\item{endpoint}{ingress endpoint. Default to "http://localhost:8080/api/v0/update"}
}
\description{
Push data points to an ingress instance.
}
