% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{to_gts_input_format}
\alias{to_gts_input_format}
\title{Convert a data frame or data table}
\usage{
to_gts_input_format(df)
}
\arguments{
\item{df}{first column must be named "timestamp".
Column names corresponding to GTS values must be in the <selector> form:
"classname\{label1=value1,label2=value2,...\}" (or "classname").
Optional geo columns must be named "<selector>.lat",
"<selector>.lon" or "<selector>.elev"}
}
\value{
data points in GTS input format as a character vector
}
\description{
Convert a data frame into GTS input format.
}
