% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_add_value.R
\name{wrp_add_value}
\alias{wrp_add_value}
\title{Add Value}
\usage{
wrp_add_value(
  wrp_con,
  tick,
  value,
  latitude = NA,
  longitude = NA,
  elevation = NA
)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{tick}{Timestamp}

\item{value}{Value of the tick.}

\item{latitude, longitude}{Coordinates.}

\item{elevation}{Elevation.}
}
\description{
The ADDVALUE function adds a value to a GTS, without checking for tick duplicates.
The added data point is appended to the GTS. The SETVALUE overrides an existing value.
}
\details{
The ADDVALUE function only works on GTS, not on lists of GTS like those generated by the use of FETCH.
If you want to use ADDVALUE after a FETCH, you will need to extract the GTS from the list,
for example by using GET with value 0 as parameter.

If latitude or longitude are NaN (not a number), value has only timestamp, elevation, value.
If elevation is NaN (not a number), value has only timestamp, latitude, longitude, value.
If elevation is NaN and longitude or latitude are NaN, value has only timestamp, value.

When adding a Geo Time Series™ or GTS Encoder value,
the value will be wrapped (as when using WRAPRAW) and added as a binary value.
}
\examples{
\dontrun{
  con <- wrp_connect()
  con %>%
    clear_script() %>%
    wrp_new_gts() %>%
    wrp_rename("test") %>%
    wrp_add_value(tick = 100, value = TRUE) %>%
    wrp_add_value(tick = 110, longitude = 45, value = TRUE) %>%
    wrp_add_value(tick = 120, latitude = 1.2, elevation = 250, value = TRUE) %>%
    wrp_add_value(tick = 130, latitude = 48.44218, longitude = -4.41427,
                  elevation = 80000, value = TRUE) %>%
    wrp_add_value(tick = 100, value = FALSE) %>%
    wrp_exec()
}

}
