% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_add_value.R
\name{wrp_add_value_df}
\alias{wrp_add_value_df}
\title{Add Value from DataFrame}
\usage{
wrp_add_value_df(
  wrp_con,
  df,
  tick = "tick",
  value = "value",
  latitude = NULL,
  longitude = NULL,
  elevation = NULL
)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{df}{A dataframe}

\item{tick}{column representing Timestamp, could be numeric or date values}

\item{value}{column representing value}

\item{latitude, longitude}{columns representing coordinates. Can be NULL.}

\item{elevation}{column representing elevation. Can be NULL.}
}
\description{
Add values to a GTS from a data.frame.
}
\examples{
\dontrun{
  library(tibble)
  df <- tribble(
    ~ tick, ~ value, ~ latitude, ~ longitude, ~ elevation,
    100   , TRUE   , NA        , NA         , NA,
    110   , TRUE   , 45        , NA         , NA,
    120   , TRUE   , NA        , 1.2        , 250,
    130   , TRUE   , 48.44218  , -4.41427   , 80000,
    100   , FALSE  , NA        , NA         , NA
  )
  con <- wrp_connect()
  con \%>\%
    clear_script() \%>\%
    wrp_new_gts() \%>\%
    wrp_rename("test") \%>\%
    wrp_add_value_df(df) \%>\%
    wrp_exec()
}

}
\seealso{
wrp_add_value
}
