% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_at_index.R
\name{wrp_at_index}
\alias{wrp_at_index}
\title{At index}
\usage{
wrp_at_index(wrp_con, index)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{index}{Index of the tick}
}
\description{
The ATINDEX function consumes a Geo Time Series™ or a list thereof from the stack,
looks at its index-th point and put on the stack a list with the timestamp, longitude,
latitude, elevation and value for the index-th point of the GTS.
}
\details{
Since 2.1, the index can be negative. In this case the effective index is index + size of the GTS.
}
\examples{

df <- data.frame(tick = seq(100, 1000, by = 100), value = 10:1)
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_rename("test") \%>\%
  wrp_relabel(list("label0", "42", "label1", "foo")) \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_at_index(4) \%>\%
  wrp_exec()
}
\references{
\url{https://www.warp10.io/doc/ATINDEX}
}
\seealso{
\code{\link[=wrp_at_tick]{wrp_at_tick()}}]
}
