% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_attributes.R
\name{wrp_attributes}
\alias{wrp_attributes}
\title{Attributes}
\usage{
wrp_attributes(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Retrieves the attributes of a Geo Time Series™.
}
\details{
The ATTRIBUTES function takes a GTS on top of the stack and push back a MAP
including all its attributes.

An attribute corresponds to a tag for a specific series. The attribute system
allows the user to add some information that can change in a series.
}
\examples{
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_rename("test") \%>\%
  wrp_set_attributes(list(attr1 = "42", attr2 = "foo")) \%>\%
  wrp_attributes() \%>\%
  wrp_exec()

}
\seealso{
\code{\link[=wrp_set_attributes]{wrp_set_attributes()}}, \code{\link[=wrp_labels]{wrp_labels()}}, \code{\link[=wrp_name]{wrp_name()}}
}
\keyword{gts}
