% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_chunk.R
\name{wrp_chunk}
\alias{wrp_chunk}
\title{Chunk}
\usage{
wrp_chunk(wrp_con, last, width, overlap, count, label, keep_empty)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{last}{The end timestamp of the most recent split to consider.}

\item{width}{The width in time units of each chunk.}

\item{overlap}{An overlap duration (in time units) between chunks,
this is useful when chunking a GTS to apply an algorithm which operates on a sliding window.}

\item{count}{The number of chunks to keep (starting from the last chunk).}

\item{label}{The label name of the label added by the split operation
(with a value corresponding to the first tick of the chunk).}

\item{keep_empty}{If true empty chunks are kept.}
}
\description{
The CHUNK is used to split a Geo Time Series™ into partial GTS.
The split operation is controlled by multiple parameters.
The series will be splitted in chunks, each chunks will form a new GTS,
with a label value corresponding to its first tick.
}
\examples{
df <- data.frame(tick = seq(100, 1100, by = 100)[-6], value = 10:1)

wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_rename("test") \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_chunk(
    last       = 1000,
    width      = 500,
    overlap    = 0,
    count      = 0,
    label      = ".chunkid",
    keep_empty = FALSE
  ) \%>\%
 wrp_exec()

}
