% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_clone.R
\name{wrp_clone}
\alias{wrp_clone}
\title{Clone}
\usage{
wrp_clone(wrp_con)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}
}
\description{
The CLONE function does a shallow copy of a given Geo Time Series™, List, Map or Vector.
Since 2.1.1 it can also clone Encoders.
}
\details{
As it makes a shallow copy, be careful when using CLONE on a List of Geo Time Series™
like those generated by the use of FETCH. Modifying a Geo Time Series™ in any of those list will alter
the matching Geo Time Series™ in the other List.
If you want to use CLONE after a FETCH, you will need to extract the Geo Time Series™ from the list,
for example by using LIST->, GET or LMAP.
}
\examples{

wrp_connect() \%>\%
  wrp_unwrap("'60VgNqxhAaCdS6_uOLtZNMGWAbGpS5xmPL4gAbC_QbCjRbBiS5KhR5KmNMGpRaJQ.NV7RqKiRqxmHLF1B23L.0N.5k..KV.N6bF.0DxeA7x..3.pV.......4EyQb2_.CJnBnBnBnGN33V.' '60VgNqxhAaCdS6_uOLtZNMGWAbGpS5xmPL4gAbC_QbCjRbBiS5KhR5KmNMGpRaJQ.NV7RqKiRqxmHLF1B27L.0N.5k..KV.N6rF.0Dxe1Bju.3.ltaOaOaOa472Jp9g0F26BnBnBnBoL007.'") \%>\% # nolint
  wrp_get(0) \%>\%
  wrp_clone() \%>\%
  wrp_add_value(tick = "NOW", value = 42) \%>\%
  set_script("[ 1 2 3 4 ]", add = "list") \%>\%
  wrp_clone() \%>\%
  set_script("42 +!", consume = "list", add = "list") \%>\%
  wrp_exec()
}
