% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_dedup.R
\name{wrp_dedup}
\alias{wrp_dedup}
\title{Dedup}
\usage{
wrp_dedup(wrp_con)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}
}
\description{
Removes duplicate ticks.
}
\details{
The DEDUP function consumes a Geo Time Series or a LIST of Geo Time Series from the top of the stack
and pushes back the Geo Time Series™ with deduplicated ticks.
}
