% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_drop.R
\name{wrp_drop}
\alias{wrp_drop}
\title{Drop}
\usage{
wrp_drop(wrp_con)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}
}
\description{
The DROP function removes the element at the top of the stack.
}
\examples{

wrp_connect() \%>\%
  set_script("foo") \%>\%
  set_script("bar") \%>\%
  wrp_drop() \%>\%
  wrp_exec()

}
