% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_dup.R
\name{wrp_dup}
\alias{wrp_dup}
\title{Dup}
\usage{
wrp_dup(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
The DUP function duplicates the element on the top of the stack, it doesn’t copy the content.
If you apply DUP to a GTS you will have two references to the same GTS object on the stack.
}
\details{
If you want to make a deep copy of a GTS, i.e. if you want copy the content
and not only the reference, you can use the CLONE function.
}
\examples{

wrp_connect() \%>\%
  set_script(123, add = "numeric") \%>\%
  wrp_dup() \%>\%
  wrp_exec()
}
