% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_fetch.R
\name{wrp_fetch}
\alias{wrp_fetch}
\title{Fetch}
\usage{
wrp_fetch(
  wrp_con,
  class = "~.*",
  labels = NULL,
  end = "NOW",
  count = 1,
  timespan = NULL
)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{class}{Selector for the GTS classes.
Can either be an exact match or a regular expression (the value should then start with a ~).}

\item{labels}{Map of labels selectors.
The key is the label name, the value a selector which can be an exact match or a regular expression
(the value should then start with a ~).}

\item{end}{Newest timestamp to consider when fetching datapoints.}

\item{count}{Maximum number of datapoints to fetch for each GTS. Incompatible with negative 'timespan'.}

\item{timespan}{Depth to consider when fetching the datapoints.
If the value is positive then it is interpreted as a duration in time units,
if it is negative then as the maximum number of datapoints to fetch.
If negative, incompatible with \code{count}.}
}
\description{
The FETCH function interacts with the Warp 10 Storage Engine to retrieve data according to given criteria.
}
\references{
\url{https://www.warp10.io/doc/FETCH}
}
\seealso{
\code{\link[=wrp_find]{wrp_find()}}
}
\keyword{gts}
