% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_find.R
\name{wrp_find}
\alias{wrp_find}
\title{Find}
\usage{
wrp_find(wrp_con, class = "~.*", labels = NULL)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{class}{Selector for the GTS classes.
Can either be an exact match or a regular expression (the value should then start with a ~).}

\item{labels}{Map of labels selectors.
The key is the label name, the value a selector which can be an exact match or a regular expression
(the value should then start with a ~).}
}
\value{
A data.frame if the request is exectued, a character string containing the Warpscript otherwise.
}
\description{
The FIND function finds Geo Time Series™ labels and attributes of GTS, not their values.
It is the fastest way to read attributes of GTS.
FIND ask only directory component of Warp 10, while FETCH first ask directory, then ask store to read GTS values.
}
