% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_find_sets.R
\name{wrp_find_sets}
\alias{wrp_find_sets}
\title{Find Sets}
\usage{
wrp_find_sets(wrp_con, class = "~.*", labels = NULL)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{class}{Selector for the GTS classes.
Can either be an exact match or a regular expression (the value should then start with a ~).}

\item{labels}{Map of labels selectors.
The key is the label name, the value a selector which can be an exact match or a regular expression
(the value should then start with a ~).}
}
\description{
The FINDSETS function returns sets of values for the classes,
labels and attributes of the Geo Time Series™ which matched the selection criteria.
}
\details{
Those results can be used for example to update a UI with lists of possible values for
labels when exploring a corpus of GTS.
}
\references{
\url{https://www.warp10.io/doc/FINDSETS}
}
\seealso{
\code{\link[=wrp_find]{wrp_find()}}, \code{\link[=wrp_find_stats]{wrp_find_stats()}}
}
\keyword{gts}
