% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_find_stats.R
\name{wrp_find_stats}
\alias{wrp_find_stats}
\title{Find stats}
\usage{
wrp_find_stats(wrp_con, class = "~.*", labels = NULL)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{class}{Selector for the GTS classes.
Can either be an exact match or a regular expression (the value should then start with a ~).}

\item{labels}{Map of labels selectors.
The key is the label name, the value a selector which can be an exact match or a regular expression
(the value should then start with a ~).}
}
\value{
The function returns a map containing information about the matching Geo Time Series.
The fields of the map are the following:
\itemize{
\item \code{gts.estimate}	Estimation of the number of matching Geo Time Series™
\item \code{classes.estimate}	Estimation of the number of distinct class names
\item \code{labelnames.estimate}	Estimation of the number of distinct label names
\item \code{labelvalues.estimate}	Estimation of the number of distinct label values
\item \code{per.class.estimate}	If the number of matching classes is below the
}

\code{directory.stats.class.maxcardinality}, this key will be associated with a map containing per
class estimate of number of GTS
\itemize{
\item \code{per.label.value.estimate}	If the number of label names in the matching GTS is below
}

\code{directory.stats.labels.maxcardinality}, this key will be associated with a map containing
an estimation of distinct label values per label name
\itemize{
\item \code{error.rate}	This is the error rate of the estimators used for computing the estimations
\item \code{partial.results}	When accessing a sharded Directory, this will be set to true if only
partial results were collected
}
}
\description{
The FINDSTATS function computes statistics on matching Geo Time Series™.
The estimations are based on the use of HyperLogLogPlus estimators.
}
\references{
\url{https://www.warp10.io/doc/FINDSTATS}
}
\seealso{
\code{\link[=wrp_find]{wrp_find()}}, \code{\link[=wrp_find_sets]{wrp_find_sets()}}
}
\keyword{gts}
