% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_first_tick.R
\name{wrp_first_tick}
\alias{wrp_first_tick}
\title{First tick}
\usage{
wrp_first_tick(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
The FIRSTTICK function pushes on the stack the timestamp of the first tick of
the Geo Time Series™ on top of the stack.
}
\details{
When applied to a list of GTS, FIRSTTICK will return the lowest first tick found across all
Geo Time Series™.
}
\examples{

df <- data.frame(tick = c(seq(100, 500, by = 100), 100, 200), value = c(10:6, 10, 9))
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_first_tick() \%>\%
  wrp_exec()
}
\references{
\url{https://www.warp10.io/doc/FIRSTTICK}
}
\seealso{
\code{\link[=wrp_at_tick]{wrp_at_tick()}}, \code{\link[=wrp_last_tick]{wrp_last_tick()}}
}
