% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_flatten.R
\name{wrp_flatten}
\alias{wrp_flatten}
\title{Flatten}
\usage{
wrp_flatten(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
The function FLATTEN inspects the top of the stack.
}
\details{
If it is a LIST of values, it inspects each value and
replaces each value which was a LIST with its content.
FLATTEN proceed recursively until all LISTs have been flattened.

If it is not a LIST, the function exits and let the stack in the same state.
}
\examples{
wrp_connect() \%>\%
  set_script("[ 'a' 'b' 'c' ]", add = "list") \%>\%
  set_script("[ 'd' 'e' [ 'f' 'g' ] ]", add = "list") \%>\%
  set_script("2 ->LIST", consume = c("list", "list"), add = "list") \%>\%
  wrp_flatten() \%>\%
  wrp_exec()
}
\references{
\url{https://www.warp10.io/doc/FLATTEN}
}
\keyword{lists}
