% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_from_b64.R
\name{wrp_from_b64}
\alias{wrp_from_b64}
\title{B64->}
\usage{
wrp_from_b64(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Decodes the base64 STRING content on top of the stack.
}
\examples{
wrp_connect() \%>\%
  set_script("aGVsbG8gd29ybGQ=") \%>\%
  wrp_from_b64() \%>\%
  wrp_from_bytes() \%>\%
  wrp_exec()
}
\seealso{
\code{wrp_from_bytes}
}
\keyword{conversion}
\keyword{strings}
