% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_from_bytes.R
\name{wrp_from_bytes}
\alias{wrp_from_bytes}
\title{BYTES->}
\usage{
wrp_from_bytes(wrp_con, charset = "UTF-8")
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{charset}{Charset to use.}
}
\description{
Converts a bytes array into a string with given a charset and put the string on top of the stack.
}
\details{
Charset could be any supported by your java platform. Example: UTF-8 US-ASCII ISO-8859-1 Cp852...

Default to UTF-8.
}
\examples{
wrp_connect() \%>\%
  set_script("aGVsbG8gd29ybGQ=") \%>\%
  wrp_from_b64() \%>\%
  wrp_from_bytes() \%>\%
  wrp_exec()
}
\seealso{
\code{wrp_from_b64}
}
\keyword{conversion}
\keyword{strings}
