% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_from_json.R
\name{wrp_from_json}
\alias{wrp_from_json}
\title{JSON->}
\usage{
wrp_from_json(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
The JSON-> function parses a string as JSON from the top of the stack and pushes the result onto the stack.
}
\examples{
wrp_connect() \%>\%
  set_script('[{"menu": {"id": "file","value": "File"}}]') \%>\%
  wrp_from_json() \%>\%
  wrp_get(0) \%>\%
  set_script('{"menu": {"id": "file","value": "File"}}') \%>\%
  wrp_from_json() \%>\%
  wrp_exec()
}
\keyword{conversion}
