% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_get.R
\name{wrp_get}
\alias{wrp_get}
\title{Get}
\usage{
wrp_get(wrp_con, index)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{index}{Depending on map construction, could be a number, a string, a list, another map, a GTS, a boolean.}
}
\description{
Retreive a value in a MAP or a list.
}
\details{
The GET function consumes on the top of the stack a list, a byte array, or a map, and the index (or the key),
then put the result on the stack.

If key is not found, it returns NULL on the stack.

If index is out of bound, GET raises an error.

Since 2.1, the index can be negative. In this case the effective index is index + size of the LIST or BYTE ARRAY.

Since 2.1, GET can operate recursively on nested lists. In this case, the index is a list.
}
\examples{
con <- wrp_connect()

con \%>\%
  set_script("{ 'foo' 42 'bar' true }", add = "map") \%>\%
  wrp_get("foo") \%>\%
  wrp_exec()

con \%>\%
  set_script("[ 3 12 15 ]", add = "list") \%>\%
  wrp_get(0) \%>\%
  wrp_exec()

con \%>\%
  set_script("{ 'foo' 42 'bar' true }", add = "map") \%>\%
  wrp_get(33) \%>\%
  wrp_exec()
}
