% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_hybridtest.R
\name{wrp_hybridtest}
\alias{wrp_hybridtest}
\title{Hybrid Test}
\usage{
wrp_hybridtest(wrp_con, period, piece, k, alpha = 0.05, ...)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{period}{Number of buckets that compose a period}

\item{piece}{Number of periods that compose a piece}

\item{k}{Upper-bound of the number of outliers to detect per piece}

\item{alpha}{Optional significance level for the statistical test. Default value is 0.05.}

\item{...}{Optional parameters of the STL call. Not used for now.}
}
\description{
The HYBRIDTEST function detects outliers in a GTS (or a LIST of GTS) which has a seasonal part.
}
\details{
Like STLESDTEST, HYBRIDTEST performs an ESDTEST onto a GTS that have been relieved of its seasonal and trend part.
But unlike the mentioned test, STL and ESDTEST are performed piecewise.
Plus, the trend is approximated with the piecewise median instead of the trend part of the STL decomposition.

A LIST of ticks (or a LIST of LIST of ticks), corresponding to the outliers, is pushed back onto the stack.

This technique was first developped at Twitter.

This function only applies to bucketized GTS of type DOUBLE.
}
