% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_interpolate.R
\name{wrp_interpolate}
\alias{wrp_interpolate}
\title{Interpolate}
\usage{
wrp_interpolate(wrp_con)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}
}
\description{
The INTERPOLATE function fills gaps in a bucketized Geo Time Series™ instance with by interpolating linearly.
}
\details{
This function has no effect on non bucketized GTS instances.
}
\examples{

df <- data.frame(tick = seq(100, 500, by = 100), value = 10:6)

wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_bucketize("mean", 500, 50, 0) \%>\%
  wrp_interpolate() \%>\%
  wrp_sort() \%>\%
  wrp_exec()

}
