% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_labels.R
\name{wrp_labels}
\alias{wrp_labels}
\title{Labels}
\usage{
wrp_labels(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
The LABELS function consumes a GTS from the stack, extracts its labels,
puts them in a map and pushes the map onto the stack.
}
\examples{
wrp_connect() \%>\%
    wrp_new_gts() \%>\%
    wrp_relabel(list(label0 = "42", label1 = "foo")) \%>\%
    wrp_labels() \%>\%
    wrp_exec()
}
\seealso{
\code{\link[=wrp_attributes]{wrp_attributes()}}, \code{\link[=wrp_name]{wrp_name()}}, \code{\link[=wrp_set_attributes]{wrp_set_attributes()}}, \code{\link[=wrp_values]{wrp_values()}}
}
\keyword{gts}
