% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_last_activity.R
\name{wrp_last_activity}
\alias{wrp_last_activity}
\title{Last Activity}
\usage{
wrp_last_activity(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Extracts the timestamp of last activity recorded in the Geo Time Series™ metadata.
}
\details{
The timestamp is expressed as platform time units elapsed since the Unix Epoch.
}
