% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_last_tick.R
\name{wrp_last_tick}
\alias{wrp_last_tick}
\title{Last tick}
\usage{
wrp_last_tick(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
The LASTTICK function pushes on the stack the newest tick of the Geo Time Series™ on top of the stack or,
if the GTS is bucketized, its last bucket.
}
\details{
When applied to a list of GTS, LASTTICK will return the greatest last tick found across all Geo Time Series™.

If the GTS does not have values, Long.MIN_VALUE is pushed.
}
\examples{

df <- data.frame(tick = c(seq(100, 500, by = 100), 100, 200), value = c(10:6, 10, 9))
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_last_tick() \%>\%
  wrp_exec()
}
\references{
\url{https://www.warp10.io/doc/LASTTICK}
}
\seealso{
\code{\link[=wrp_at_tick]{wrp_at_tick()}}, \code{\link[=wrp_first_tick]{wrp_first_tick()}}
}
