% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_merge.R
\name{wrp_merge}
\alias{wrp_merge}
\title{Merge}
\usage{
wrp_merge(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Merge several Geo Time Series™ together.
}
\details{
The MERGE function expects a LIST of Geo Time Series™ on the top of the stack.
It will consume these GTS and push onto the stack a single GTS instance with all measurements found in the GTS
instances present in the LIST.
The name and labels of the resulting GTS instance are those of the first one of the LIST.
}
\examples{
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_rename("a") \%>\%
  wrp_relabel(label1 = "foo") \%>\%
  wrp_add_value_df(data.frame(tick = 1:5, value = c(1, 1, 1, 2, 2))) \%>\%
  wrp_new_gts() \%>\% wrp_rename("b") \%>\%
  wrp_relabel(label3 = "bar") \%>\%
  wrp_add_value_df(data.frame(tick = c(1, 7:10), value = c(2, 3, 4, 4, 4))) \%>\%
  wrp_merge() \%>\%
  wrp_exec()
}
\keyword{gts}
