% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_meta.R
\name{wrp_meta}
\alias{wrp_meta}
\title{Meta}
\usage{
wrp_meta(wrp_con, write_token = Sys.getenv("ML_WARP10_WRITE_TOKEN"))
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{write_token}{GTS list with new or modified attributes}
}
\description{
Stores the attributes of a list of Geo Time Series™ in the Warp 10 backend.
}
\details{
This function expects on top of the stack a write token which will be used
to authenticate with the Warp 10 backend and a Geo Time Series™ or a list
thereof.

Every Geo Time Series™ to which META is applied must have a non empty name
and attributes (possibly empty).
}
\seealso{
\code{\link[=wrp_set_attributes]{wrp_set_attributes()}}, \code{\link[=wrp_update]{wrp_update()}}
}
\keyword{gts}
