% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_name.R
\name{wrp_name}
\alias{wrp_name}
\title{Name}
\usage{
wrp_name(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
The NAME function consumes a Geo Time Series™ from the stack,
extracts its class name, and pushes it onto the stack.
}
\examples{
wrp_connect() \%>\%
    wrp_new_gts() \%>\%
    wrp_rename("GTS1") \%>\%
    wrp_name() \%>\%
    wrp_exec()
}
\seealso{
\code{\link[=wrp_attributes]{wrp_attributes()}}, \code{\link[=wrp_labels]{wrp_labels()}}, \code{\link[=wrp_rename]{wrp_rename()}}, \code{\link[=wrp_set_attributes]{wrp_set_attributes()}}, \code{\link[=wrp_values]{wrp_values()}}
}
\keyword{gts}
