% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_new_gts.R
\name{wrp_new_gts}
\alias{wrp_new_gts}
\title{New GTS}
\usage{
wrp_new_gts(wrp_con)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}
}
\description{
The NEWGTS function pushes onto the stack an empty Geo Time Series instance.
}
\examples{
\dontrun{
  con <- wrp_connect()
  con %>%
    clear_script() %>%
    wrp_new_gts()
}

}
