% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_parse_selector.R
\name{wrp_parse_selector}
\alias{wrp_parse_selector}
\title{Parse Selector}
\usage{
wrp_parse_selector(wrp_con, selector)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{selector}{String selector.}
}
\description{
The PARSESELECTOR function parses a GTS Selector (string) and pushes the class selector and
labels selector on the stack.
}
\details{
Combined with \code{\link[=wrp_to_selector]{wrp_to_selector()}}, it can be used to build a selector from a subset of GTS.
}
\examples{
wrp_connect() \%>\%
  wrp_parse_selector('io.senx.tutorial.sensors.temperature{sensorId=01,sensortype~numeric.*}') \%>\%
  wrp_exec()

}
