% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_reduce.R
\name{wrp_reduce}
\alias{wrp_reduce}
\title{Reduce}
\usage{
wrp_reduce(wrp_con, reducer, labels = NULL, override_ticks = FALSE)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{reducer}{Reducer function to apply.}

\item{labels}{List of label names to consider for creating equivalence
classes. If the list is empty all Geo Time Series™ will end up in the same
equivalence class. If the labels parameter NULL, all labels will be
considered.}

\item{override_ticks}{Boolean, if true allows the reducer to modify the ticks
of the GTSs. If not set, the reducer cannot modify the ticks.}
}
\description{
The REDUCE framework groups Geo Time Series™ by equivalence classes based on
label values and applies a function on each equivalence class of
Geo Time Series™, considering the values of each tick and producing one GTS
per equivalence class.
}
\details{
As the selected reducer function is applied tick by tick, it is usually wise
to BUCKETIZE the GTS first with a common value of lastbucket.

The labels whose values are common to all the GTS in an equivalence class
will be retained.

Since 2.1 you can make REDUCE override the GTSs ticks. This usage in mainly
to be used in conjuction with MACROREDUCER.
}
\references{
\url{https://www.warp10.io/doc/REDUCE}
}
\keyword{framework}
\keyword{reducer,}
