% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_remove_tick.R
\name{wrp_remove_tick}
\alias{wrp_remove_tick}
\title{Remove tick}
\usage{
wrp_remove_tick(wrp_con, tick = NULL)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{tick}{Tick (timestamp) to remove.
If NULL, takes element on the stack.}
}
\description{
The REMOVETICK function transforms a Geo Time Series™ by removing all occurrences of a tick or
ticks.
}
\examples{

wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_rename("test") \%>\%
  wrp_add_value(0, 0) \%>\%
  wrp_add_value(1, 1) \%>\%
  wrp_add_value(2, 2) \%>\%
  wrp_remove_tick(1) \%>\%
  wrp_exec()

}
