% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_rename.R
\name{wrp_rename}
\alias{wrp_rename}
\title{Rename}
\usage{
wrp_rename(wrp_con, name)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{name}{New name to give to the GTS.}
}
\description{
The RENAME function changes the name of a Geo Time Series™, an Encoder or a list thereof.
This operation may be required prior to using some functions.
If the name starts with '+', the specified name will be appended to the current name of the GTS or Encoder.
}
\details{
If you want to rename a Geo Time Series or an Encoder so its name starts with a '+',
you must first set its name to the empty string then use rename with the desired name prefixed
with an additional '+'.
}
\examples{
\dontrun{
  con <- wrp_connect()
  con \%>\%
    clear_script() \%>\%
    wrp_newgts() \%>\%
    wrp_rename("foo") \%>\%
    wrp_rename("+bar") \%>\%
    wrp_exec()

  con \%>\%
    clear_script() \%>\%
    wrp_newgts() \%>\%
    wrp_rename("foo") \%>\%
    wrp_rename("") \%>\%
    wrp_rename("++bar") \%>\%
    wrp_exec()
}

}
