% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_reverse.R
\name{wrp_reverse}
\alias{wrp_reverse}
\title{Reverse}
\usage{
wrp_reverse(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Reverses the order of the elements of the list or the string.
}
\details{
The REVERSE function inverts the order of the elements of the list or the
string on the top of the stack. Beware, REVERSE do not create a new object.
Use CLONEREVERSE if you want to keep input.

Applied on a byte array, REVERSE reverses the endianness.
}
\examples{
wrp_connect() \%>\%
  set_script(c("eins", "zwei", "drei")) \%>\%
  wrp_store("mylist") \%>\%
  wrp_reverse() \%>\%
  set_script("swap tac") \%>\%
  wrp_store("mystring") \%>\%
  wrp_reverse() \%>\%
  wrp_exec()

}
\keyword{binary}
\keyword{lists,}
\keyword{strings,}
