% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_set_attributes.R
\name{wrp_set_attributes}
\alias{wrp_set_attributes}
\title{Set attributes}
\usage{
wrp_set_attributes(wrp_con, ...)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{...}{Named parameters of attributes names to values.}
}
\description{
Modifies the attributes of a Geo Time Series™, an encoder or a list thereof.
The SETATTRIBUTES function expects a parameter MAP whose entries are attributes to set or alter.
}
\details{
If the parameter MAP has an entry with a NULL key,
the SETATTRIBUTES function will use the rest of the MAP as the actual attributes to set for the GTS.
If no NULL key exist, then the other entries of the MAP will alter the existing attributes.

An entry with an empty STRING value will have the effect of removing the
attribute from the attributes of the GTSs or encoders.
}
\examples{
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_clone() \%>\%
  wrp_set_attributes(foo = "bar", bar = "foo") \%>\%
  wrp_clone() \%>\%
  wrp_set_attributes("null" = NULL, star = 'treck') \%>\%
  wrp_clone() \%>\%
  wrp_set_attributes('next' = 'generation', heckle = 'jeckle') \%>\%
  wrp_clone() \%>\%
  wrp_set_attributes('next' = '') \%>\%
  wrp_clone() \%>\%
  wrp_set_attributes(heckle = 'Peter') \%>\%
  wrp_exec()

}
\seealso{
\code{\link[=wrp_name]{wrp_name()}}, \code{\link[=wrp_relabel]{wrp_relabel()}}, \code{\link[=wrp_rename]{wrp_rename()}}, \code{\link[=wrp_meta]{wrp_meta()}}, \code{\link[=wrp_attributes]{wrp_attributes()}}, \code{\link[=wrp_labels]{wrp_labels()}}
}
\keyword{encoder}
\keyword{gts}
