% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_size.R
\name{wrp_size}
\alias{wrp_size}
\title{Size}
\usage{
wrp_size(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Returns the size of the input parameter.
}
\details{
The SIZE function computes the size of a LIST, MAP, GTS or ENCODER (number of values), STRING,
byte array of GEOSHAPE (number of cells).
}
\examples{
df <- data.frame(tick = seq(100, 1000, by = 100), value = 10:1)
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_rename("test name") \%>\%
  wrp_relabel(list("label0", "42", "label1", "foo")) \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_clone() \%>\%
  wrp_size() \%>\%
  set_script("{ 'label0' '42' 'label1' 'foo' }", add = "map") \%>\%
  wrp_dup() \%>\%
  wrp_size() \%>\%
  set_script("[ 'label0' '42' 'label1' 'foo' ]", add = "list") \%>\%
  wrp_dup() \%>\%
  set_script("one \%25") \%>\%
  wrp_dup() \%>\%
  wrp_size() \%>\%
  wrp_exec()

}
\references{
\url{https://www.warp10.io/doc/SIZE}
}
\seealso{
\code{\link[=wrp_new_gts]{wrp_new_gts()}}
}
\keyword{geo}
\keyword{gts}
\keyword{lists}
\keyword{maps}
\keyword{strings}
