% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_sort.R
\name{wrp_sort}
\alias{wrp_sort}
\title{Sort}
\usage{
wrp_sort(wrp_con)
}
\description{
Sorts a Geo Time Series™ or each GTS of a list of GTS in ascending tick order.
}
\examples{

df <- data.frame(tick = 10:1, value = runif(10))

wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_sort() \%>\%
  wrp_exec()

}
