% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_store.R
\name{wrp_store}
\alias{wrp_store}
\title{Store}
\usage{
wrp_store(wrp_con, symbol)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{symbol}{Name of the symbol to modify.}
}
\description{
The STORE function stores a value in a symbol.
}
\examples{
wrp_connect() \%>\%
  set_script("42", add = "long") \%>\%
  wrp_store("foo") \%>\%
  wrp_exec()
}
