% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_swap.R
\name{wrp_swap}
\alias{wrp_swap}
\title{Swap}
\usage{
wrp_swap(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Exchanges the positions of the top two elements of the stack.
}
\examples{
wrp_connect() \%>\%
    set_script("Thrid level") \%>\%
    set_script("Second level") \%>\%
    set_script("Top") \%>\%
    wrp_swap() \%>\%
    wrp_exec()
}
\keyword{stack}
