% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_tick_list.R
\name{wrp_tick_list}
\alias{wrp_tick_list}
\title{Tick list}
\usage{
wrp_tick_list(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Pushes onto the stack the list of ticks of a Geo Time Series™ or encoder.
The ticks appear in the order in which they are found.
}
\details{
If a list of Geo Time Series™ is given,
the result is a list of list of ticks as if the TICKLIST function was applied to each
Geo Time Series™ separately.
}
\examples{

df <- data.frame(tick = 1:10, value = rnorm(10))
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_tick_list() \%>\%
  wrp_exec()
}
\seealso{
\code{\link[=wrp_ticks]{wrp_ticks()}}
}
