% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_ticks.R
\name{wrp_ticks}
\alias{wrp_ticks}
\title{Ticks}
\usage{
wrp_ticks(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Pushes onto the stack the sorted list of distinct ticks of a Geo Time Series™.
If a list of Geo Time Series™ is given,
all the distinct ticks are returned in a single sorted list as if all Geo Time Series™ were merged.
}
\details{
This function is not to be confused with \code{\link[=wrp_tick_list]{wrp_tick_list()}} as this function returns distinct ticks and
acts as it merges Geo Time Series™.
}
\examples{
df <- data.frame(tick = 1:10, value = rnorm(10))
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_ticks() \%>\%
  wrp_exec()
}
\references{
\url{https://www.warp10.io/doc/TICKS}
}
\seealso{
\code{\link[=wrp_tick_list]{wrp_tick_list()}}
}
\keyword{gts}
