% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_to_selector.R
\name{wrp_to_selector}
\alias{wrp_to_selector}
\title{To selector}
\usage{
wrp_to_selector(wrp_con)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}
}
\description{
The TOSELECTOR function takes a Geo Time Series™ or Geo Time Series™ LIST from the top of the stack and,
for each encountered GTS, replace it with a selector which would select it.
}
\details{
This selector can be used as input of PARSESELECTOR for a FETCH.
}
\examples{

wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_rename("test name") \%>\%
  wrp_relabel(c("label0", 42, "label1", "foo")) \%>\%
  wrp_add_value(100, value = 10) \%>\%
  wrp_add_value(200, value = 9) \%>\%
  wrp_add_value(300, value = 8) \%>\%
  wrp_store("gts1") \%>\%
  set_script("$gts1", add = "gts") \%>\%
  wrp_clone() \%>\%
  wrp_relabel(c("label0", "33")) \%>\%
  wrp_store("gts2") \%>\%
  wrp_drop() \%>\%
  set_script("[ $gts1 $gts2 ]", add = "gtslist") \%>\%
  wrp_to_selector()
}
