% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_unbucketize.R
\name{wrp_unbucketize}
\alias{wrp_unbucketize}
\title{Unbucketize}
\usage{
wrp_unbucketize(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Transforms a bucketized Geo Time Series™ into a non bucketized one.
Note that this function transforms the original GTS, it does not clone it.
}
\examples{
df <- data.frame(tick = 1:100, value = TRUE)
wrp_connect() \%>\%
wrp_new_gts() \%>\%
wrp_add_value_df(df) \%>\%
wrp_bucketize("last", span = 10) \%>\%
wrp_get(0) \%>\%
wrp_unbucketize() \%>\%
wrp_exec()

}
\references{
\url{https://www.warp10.io/doc/UNBUCKETIZE}
}
\seealso{
\code{\link[=wrp_bucketize]{wrp_bucketize()}}
}
\keyword{gts}
