% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_unwrap.R
\name{wrp_unwrap}
\alias{wrp_unwrap}
\title{Unwrap}
\usage{
wrp_unwrap(wrp_con)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}
}
\description{
Unwraps packed Geo Time Series™ instances.
}
\examples{
wrp_connect() \%>\%
  set_script("60V.5k.L.0N.5k..KV.N5GyA1.........0nNL7O4W..rXE6gwV....Lm.3G..") \%>\%
  wrp_unwrap() \%>\%
  wrp_exec()
}
