% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_update.R
\name{wrp_update}
\alias{wrp_update}
\title{Update}
\usage{
wrp_update(wrp_con, token = Sys.getenv("ML_WARP10_WRITE_TOKEN"))
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}

\item{token}{A valid token used to write GTS}
}
\description{
Pushes Geo Time Series™ data to the Warp 10 instance.
}
\details{
The UPDATE function allows you to push data directly from your WarpScript code
without having to retrieve the data and use the /update endpoint.

The GTS or Encoder instances to push MUST have a non empty name and in the
case of Geo Time Series™ MUST have been renamed (to avoid pushing data
by mistake which could override existing data you just retrieved).
}
\seealso{
\code{\link[=wrp_meta]{wrp_meta()}}
}
