% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_values.R
\name{wrp_values}
\alias{wrp_values}
\title{Values}
\usage{
wrp_values(wrp_con)
}
\arguments{
\item{wrp_con}{\link{wrp_connect} object.}
}
\description{
Gets the values of Geo Time Series™ or encoders.
}
\details{
The VALUES function consumes a Geo Time Series™, an encoder or a list thereof
from the stack, and it replaces each instance of Geo Time Series™ or encoder
by a list of its values.
}
\examples{
df <- data.frame(tick = c(100, 200, 300, 400, 500, 700, 800, 900, 1000, 1100), value = 10:1)
wrp_connect() \%>\%
  wrp_new_gts() \%>\%
  wrp_add_value_df(df) \%>\%
  wrp_values() \%>\%
  wrp_exec()

}
\seealso{
\code{\link[=wrp_labels]{wrp_labels()}}
}
\keyword{gts}
