% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts.R
\name{as_gts}
\alias{as_gts}
\title{Coersce object to GTS tibbles}
\usage{
as_gts(x, class = "", labels = list(), combine = TRUE, .funs = "first")
}
\arguments{
\item{x}{An object that could reasoably be coerced to a gts tibble.}

\item{class}{Selector for the GTS classes.
Can either be an exact match or a regular expression (the value should then start with a ~).}

\item{labels}{Map of labels selectors.
The key is the label name, the value a selector which can be an exact match or a regular expression
(the value should then start with a ~).}

\item{combine}{If TRUE, combine multiple time series into one by grouping dates.}

\item{.funs}{A function \code{fun}, a quosure style lambda \code{~ fun(.)} or a list of either form.}
}
\description{
`as_gts()` turns an existing object, such as a data frame into
an extended tibble with class ['gts'], and attributes corresponding
to the GTS attributes + class.
A [`GTS`] object is basically a data frame with some information about the
timeserie.
The resulting tibble contains at least two columns: `timestamp` and `value`.

`as_gts()` is an S3 generic.
}
\examples{

x <- data.frame(timestamp = 1:10, value = rnorm(10))
as_gts(x)
}
