% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_endpoint}
\alias{assert_endpoint}
\alias{assert_token}
\title{Helper function}
\usage{
assert_endpoint(endpoint)

assert_token(token)
}
\arguments{
\item{endpoint}{}

\item{token}{}
}
\description{
Helper function to assert that values are correctly sets.
}
