% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{build_res}
\alias{build_res}
\alias{build_res.data}
\alias{build_res.ldata}
\alias{build_res.default}
\alias{build_res.map}
\alias{build_res.lastactivity}
\alias{build_res.list}
\alias{build_res.gts}
\title{Build results}
\usage{
build_res(object, data, combine, .funs)

\method{build_res}{data}(object, data, ...)

\method{build_res}{ldata}(object, data, ...)

\method{build_res}{default}(object, data, ...)

\method{build_res}{map}(object, data, ...)

\method{build_res}{lastactivity}(object, data, ...)

\method{build_res}{list}(object, data, ...)

\method{build_res}{gts}(object, data, combine, .funs)
}
\arguments{
\item{object}{A string corresponding to the object fetched from Warp10 database.}

\item{data}{A list resulting of a parsed json of all results.}

\item{combine}{If TRUE, combine multiple time series into one by grouping dates.}

\item{.funs}{A function \code{fun}, a quosure style lambda \code{~ fun(.)} or a list of either form.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
Build results from parsed json file.
}
