% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{build_res}
\alias{build_res}
\alias{build_res.default}
\alias{build_res.gts}
\title{Build results}
\usage{
build_res(object, data, combine, operator)

\method{build_res}{default}(object, data, ...)

\method{build_res}{gts}(object, data, combine, operator)
}
\arguments{
\item{object}{A string corresponding to the object fetched from Warp10 database.}

\item{data}{A list resulting of a parsed json of all results.}

\item{combine}{For LGTS, should all GTS be combined into one single GTS.}

\item{operator}{If `combine=TRUE` which function should be used to combine all the elements.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
Build results from parsed json file.
}
