% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{build_res}
\alias{build_res}
\title{Build results}
\usage{
build_res(object, data)
}
\arguments{
\item{object}{A string corresponding to the object fetched from Warp10 database.}

\item{data}{A list resulting of a parsed json of all results.}
}
\description{
Build results from parsed json file.
}
