% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connexion.R
\name{wrp_connect}
\alias{wrp_connect}
\title{Create Connection}
\usage{
wrp_connect(endpoint = get_endpoint(), token = get_token(endpoint))
}
\arguments{
\item{endpoint}{egress endpoint.}

\item{token}{a token needed to connect to database}
}
\description{
Create a connection to a Warp 10 database.
The connexion is an R6 object with few informations
}
