% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{wrp_exec}
\alias{wrp_exec}
\title{Execute Warp Script}
\usage{
wrp_exec(wrp_con, combine = TRUE, .funs = "sum")
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{combine}{If TRUE, combine multiple time series into one by grouping dates.}

\item{.funs}{A function \code{fun}, a quosure style lambda \code{~ fun(.)} or a list of either form.}
}
\description{
Execute the Warp Script by sending it to the server.
}
