% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_fetch.R
\name{wrp_fetch}
\alias{wrp_fetch}
\title{Fetch}
\usage{
wrp_fetch(
  wrp_con,
  token = get_token(),
  class = "~.*",
  labels = NULL,
  end = "NOW",
  stop = -1
)
}
\arguments{
\item{wrp_con}{[wrp_connect] object.}

\item{token}{a token needed to connect to database}

\item{class}{Selector for the GTS classes.
Can either be an exact match or a regular expression (the value should then start with a ~).}

\item{labels}{Map of labels selectors.
The key is the label name, the value a selector which can be an exact match or a regular expression
(the value should then start with a ~).}

\item{end}{Most recent timestamp to consider when fetching datapoints. Could be a Posixct Date.}

\item{stop}{Either a number, a date or a duration, see details.}
}
\description{
The FETCH function interacts with the Warp 10 Storage Engine to retrieve data according to given criteria.
}
\details{
The stop parameter can either be a number or a date.
If a number, the depth to consider when fetching datapoints,
if the value is positive then it is interpreted as a duration,
if it is negative then as the maximum number of datapoints to fetch.
If a date, the stop timestamp to consider when fetching data, in ISO-8601 format.
If a duration, the stop timestamp will be computed first before the request is adressed to Warp 10.
}
